tests/wpt/web-platform-tests/css/css-ui/outline-018.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Basic User Interface Test: non interpolable outline-style animation</title>
<link rel="author" title="Florian Rivoal" href="http://florian.rivoal.net">
<link rel="help" href="https://drafts.csswg.org/css-ui-3/#propdef-outline-style">
<link rel="help" href="https://www.w3.org/TR/web-animations-1/#dom-animatable-animate">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
@keyframes outline-anim {
  from {
    outline: solid 1px black;
  }
  to {
    outline: dotted 1px black;
  }
}

#test {
  animation: outline-anim 3s -1.5s paused linear;
  outline: solid 1px black;
}
</style>
<body>
  <div id="test"></div>
  <div id=log></div>

  <script>
    test(
      function(){
        var test = document.getElementById("test");

        test.style.animation = "outline-anim 3s 0s paused linear"
        assert_equals(getComputedStyle(test).outlineStyle, 'solid');
        test.style.animation = "outline-anim 3s -1s paused linear"
        assert_equals(getComputedStyle(test).outlineStyle, 'solid');
        test.style.animation = "outline-anim 3s -2s paused linear"
        assert_equals(getComputedStyle(test).outlineStyle, 'dotted');
        test.style.animation = "outline-anim 3s 0s paused reverse-linear"
        assert_equals(getComputedStyle(test).outlineStyle, 'dotted');
      }, "outline-style is animated as a discrete type");
</script>
</body>
