LayoutTests/imported/w3c/web-platform-tests/css/css-ui/outline-018-expected.txt


PASS outline-style is animated as a discrete type
